requirements.txt
# required to serve the model
Flask==1.1.1
gunicorn
gevent

# dependencies for the custom model
numpy==1.18.1
#pandas==1.0.1
#requests==2.22.0
#scikit-learn==0.22.1
#tensorflow==1.15
tensorflow-gpu==2.0
tensorflow==2.0
#tqdm==4.43.0
#urllib3==1.25.8
opencv-python==4.2.0.32
keras-retinanet==0.5.1
Pillow==7.0.0
keras==2.2.5